move lattice code
